nlp-practice/text-similarity-comparison/requirements.txt
absl-py==0.5.0
astor==0.7.1
boto==2.49.0
boto3==1.9.19
botocore==1.12.19
bz2file==0.98
certifi==2018.8.24
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.0.2
Flask-Environments==0.1
gast==0.2.0
gensim==3.6.0
grpcio==1.15.0
h5py==2.8.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
Keras==2.1.5
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.0
mecab-python-windows==0.996.1
numpy==1.15.2
pandas==0.23.4
protobuf==3.6.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
smart-open==1.7.1
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.14.1
pillow==5.3.0
matplotlib==3.0.1